requirements.txt
cffi==1.17.1
cryptography==45.0.6
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
pycparser==2.22
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.1.1
PyYAML==6.0.2
